Fix URL

+++ src/main/java/io/trunk/jenkins/model/Timestamp.java
@@ -16,5 +16,5 @@ public interface Timestamp {
 
     long seconds();
 
-    long nanos();
+    int nanos();
 }
